Selections

Description

These functions are constructed as described in paper Exhaustible Sets in Higher-Type Computation, Logical Methods in Computer Science, 2008, by Martin Escardo.

Type definitions

type Nat = Int

type Baire = Nat -> Nat

type J t = (t -> Bool) -> t

type K t = (t -> Bool) -> Bool

Selection functions

findN :: K Nat -> J Nat

Mapping of existential quantifier for a subset of Nat to its selection function.

findNN :: K Baire -> J Baire

Mapping of existential quantifier for a subset of Baire to its selection function.

memoNN :: K Baire -> J Baire

Function findNN implemented with memoization.

findNNN :: K (Baire -> Nat) -> J (Baire -> Nat)

Mapping of existential quantifier for a subset of Baire -> Nat to its selection function.