Tychonoff

Description

In this module we define a selection function on a product of compact spaces with selection functions, which is also compact according to Tychonoff theorem. Code is described in paper What Sequential Games, the Tychonoff Theorem and the Double-Negation Shift have in Common, 2010, by Martin Escardo.

Documentation

findCantor :: J Bool [Bool]

Selection function on Cantor space.

image :: (x -> y) -> J r x -> J r y

Functor of a monad J r.

code :: (Integer -> Bool) -> [Bool]

Function that codes functions Integer -> Bool as boolean lists.

decode :: [Bool] -> Integer -> Bool

Inverse of function code.

at :: [x] -> Integer -> x

Call at xs n returns (n+1)-th element of a list xs.

findFunction :: J Bool (Integer -> Bool)

Selection function of functions Integer -> Bool.

forsomeFunction :: K Bool (Integer -> Bool)

Existential quantifier over functions Integer -> Bool.

foreveryFunction :: K Bool (Integer -> Bool)

Universal quantifier over functions Integer -> Bool.

equal :: Eq z => ((Integer -> Bool) -> z) -> ((Integer -> Bool) -> z) -> Bool

Function that checks equality of functionals (Integer -> Bool) -> z.

coerce :: Bool -> Integer

Function that maps True to 1 and False to 0.

f :: (Integer -> Bool) -> Integer

Examples of functional useful for testing function equal.

h :: (Integer -> Bool) -> Integer

Examples of functional useful for testing function equal.

g :: (Integer -> Bool) -> Integer

Examples of functional useful for testing function equal.

fan :: Eq z => ([Bool] -> z) -> Int

Calculates modulus of uniform continuity of given functional.

least :: (Int -> Bool) -> Int

Call least p returns the smallest natural number n, such that p n = True.

(-->) :: Bool -> Bool -> Bool

Logical implication.

agree :: Int -> [Bool] -> [Bool] -> Bool

Call agree n as bs checks if sequences as and bs agree on the first n terms.

forsome :: K Bool [Bool]

Existential and universal quantifier over Cantor space.

forevery :: K Bool [Bool]

Existential and universal quantifier over Cantor space.